Preliminary file restructuring
